requirements.txt
# Web Scraping
scrapy>=2.11.0
scrapy-user-agents>=0.1.1
requests>=2.31.0
googlemaps>=4.10.0

# Database
pymongo>=4.6.0

# Dashboard
streamlit>=1.28.1
plotly>=5.17.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Data Processing & NLP
pandas>=2.1.0
numpy>=1.24.0
nltk>=3.8.1
textblob>=0.17.1
wordcloud>=1.9.2
scikit-learn>=1.3.0

# Scheduler
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0

# Geospatial
geopy>=2.4.0

# Image processing for charts
pillow>=10.0.0

# Additional dependencies
matplotlib>=3.7.0